requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
itsdangerous==2.2.0
bcrypt==4.2.0
sqlalchemy==2.0.35
playwright==1.47.0
apscheduler==3.10.4
geopy==2.4.1
folium==0.17.0
pyyaml==6.0.2
python-dotenv==1.0.1
httpx==0.27.2
sse-starlette==2.1.3
beautifulsoup4==4.12.3
lxml==5.2.2
